Moved Validator to the androidshared module
