+++ androidshared/src/test/java/org/odk/collect/androidshared/utils/ValidatorTest.kt
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.odk.collect.shared
+package org.odk.collect.androidshared.utils
 
 import junit.framework.TestCase.assertFalse
 import junit.framework.TestCase.assertTrue
 import org.junit.Test
-import org.odk.collect.shared.strings.Validator
 
 class ValidatorTest {
     @Test
